implements full memory mapped interval stabbing
