+++ src/main.cpp
@@ -20,7 +20,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 #include <iostream>
 #include <vector>
 #include <random>
-#include "sintervalstab.hpp"
+#include "mmintervalstab.hpp"
 #include "args.hxx"
 
 using namespace intervalstab;
